pyproject.toml
[project]
name = "greencreditai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite>=0.22.1",
    "bcrypt==4.0.1",
    "certifi>=2026.1.4",
    "dashscope>=1.25.8",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "langchain-qdrant>=1.1.0",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.6",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "loguru>=0.7.3",
    "networkx>=3.6.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pypdf>=6.6.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "python-pptx>=1.0.2",
    "qdrant-client>=1.16.2",
    "sqlmodel>=0.0.31",
    "tavily-python>=0.7.19",
    "unstructured>=0.18.27",
    "urllib3<2.0.0",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true


[tool.ruff]
target-version = "py312"

line-length = 120

[tool.ruff.lint]
# 2. 启用规则集
select = [
    "F",    # Pyflakes: 发现逻辑错误 (如未使用的变量)
    "E",    # pycodestyle errors: 代码风格错误
    "W",    # pycodestyle warnings: 代码风格警告
    "I",    # isort: 自动对 import 进行排序和归类 (非常推荐!)
    "UP",   # pyupgrade: 自动升级语法 (包含你想要的 X | Y)
    "B",    # flake8-bugbear: 发现潜在的 Bug 设计
    "SIM",  # flake8-simplify: 建议更简单的代码写法
    "N",    # pep8-naming: 检查变量/函数命名规范
]

# 忽略的规则
ignore = [
    "E501", # 忽略行过长检查 (交给 format 自动处理，或者容忍特例)
    "B008", # 忽略 FastAPI Depends 的默认参数检查
]

# Exclude typically ignored directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pantheon.github.io",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
